// frontend/src/api/generated/models/FlashcardUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Flashcard update schema.
 */
export type FlashcardUpdate = {
    /**
     * Front of card
     */
    front_text?: (string | null);
    /**
     * Back of card
     */
    back_text?: (string | null);
    /**
     * Front media URL
     */
    front_media_url?: (string | null);
    /**
     * Back media URL
     */
    back_media_url?: (string | null);
};